refactor(i18n): move config-context.ts to components/ per convention
Quality cleanup (/simplify): context/injection-key files live in src/components/
in this repo (toast-context, radio-context, tabs-context, …) per CLAUDE.md.
Moved config-context.ts there to match the established useToast/toast-context
layering and updated the two imports. No behavior change.
